+++ app/src/main/java/com/ab/hicarerun/network/models/WalletModel/WalletData.kt
@@ -7,7 +7,7 @@ data class WalletData (
     @SerializedName("burnedList") val burnedList : List<BurnedList>?,
     @SerializedName("pointsInWallet") val pointsInWallet : Int?,
     @SerializedName("totalReedemablePercentageInWallet") val totalReedemablePercentageInWallet : Int?,
-    @SerializedName("totalRedeemablePointsInWallet") val totalRedeemablePointsInWallet : Int?,
+    @SerializedName("totalRedeemablePointsInWallet") val totalRedeemablePointsInWallet : Double?,
     @SerializedName("pointsEarned") val pointsEarned : Int?,
     @SerializedName("pointsBurned") val pointsBurned : Int?,
     @SerializedName("historyList") val historyList : List<HistoryList>?
